+++ amqp/public_include/fty/messagebus2/amqp/MessageBusAmqp.h
@@ -27,40 +27,41 @@ namespace fty::messagebus2::amqp {
 
 // Default amqp end point
 static auto constexpr DEFAULT_ENDPOINT{"amqp://127.0.0.1:5672"};
-
 static auto constexpr BUS_IDENTITY{"AMQP"};
-static const std::string TOPIC_PREFIX = "topic://";
-static const std::string QUEUE_PREFIX = "queue://";
 
-class MsgBusAmqp;
+class AmqpClient;
 
 class MessageBusAmqp final : public MessageBus
 {
 public:
-    MessageBusAmqp(const ClientName& clientName = utils::getClientId("MessageBusAmqp"), const Endpoint& endpoint = DEFAULT_ENDPOINT);
-
+    MessageBusAmqp(const std::string& clientName = utils::getClientId("MessageBusAmqp"), const Endpoint& endpoint = DEFAULT_ENDPOINT);
     ~MessageBusAmqp() = default;
 
     MessageBusAmqp(MessageBusAmqp&&) = delete;
-    MessageBusAmqp& operator=(MessageBusAmqp&&) = delete;
-    MessageBusAmqp(const MessageBusAmqp&)       = delete;
-    MessageBusAmqp& operator=(const MessageBusAmqp&) = delete;
+    MessageBusAmqp& operator = (MessageBusAmqp&&) = delete;
+    MessageBusAmqp(const MessageBusAmqp&) = delete;
+    MessageBusAmqp& operator = (const MessageBusAmqp&) = delete;
 
     [[nodiscard]] fty::Expected<void, ComState>      connect() noexcept override;
     [[nodiscard]] fty::Expected<void, DeliveryState> send(const Message& msg) noexcept override;
     [[nodiscard]] fty::Expected<void, DeliveryState> receive(
-        const Address& address, MessageListener&& func, const std::string& filter = {}) noexcept override;
-    [[nodiscard]] fty::Expected<void, DeliveryState>    unreceive(const Address& address) noexcept override;
-    [[nodiscard]] fty::Expected<Message, DeliveryState> request(const Message& msg, int timeOut) noexcept override;
-
-    void setConnectionErrorListener(ConnectionErrorListener errorListener);
-
+        const Address& address, MessageListener&& messageListener, const std::string& filter = {}) noexcept override;
+    [[nodiscard]] fty::Expected<void, DeliveryState> unreceive(const Address& address, const std::string& filter = {}) noexcept override;
+    // Sync request with timeout
+    [[nodiscard]] fty::Expected<Message, DeliveryState> request(const Message& message, int timeoutInSeconds) noexcept override;
     [[nodiscard]] const ClientName& clientName() const noexcept override;
-    [[nodiscard]] const Identity&   identity() const noexcept override;
+    [[nodiscard]] const Identity& identity() const noexcept override;
 
-    using MessageBus::receive;
 private:
-    std::shared_ptr<MsgBusAmqp> m_busAmqp;
+    // Test if the service is available or not
+    bool isServiceAvailable();
+
+    // Client name
+    std::string m_clientName{};
+    // Amqp endpoint
+    Endpoint    m_endpoint{};
+    // AmqpClient instance
+    std::shared_ptr<AmqpClient> m_clientPtr;
 };
 
 } // namespace fty::messagebus2::amqp