+++ amqp/src/AmqpClient.cpp
@@ -50,7 +50,7 @@ namespace fty::messagebus2::amqp {
 using namespace fty::messagebus2;
 using MessageListener = fty::messagebus2::MessageListener;
 
-static auto constexpr TIMEOUT = std::chrono::seconds(5);
+static auto constexpr TIMEOUT_MS = 5000;
 
 AmqpClient::AmqpClient(const Endpoint& url)
     : m_url(url), m_pool(std::make_shared<fty::messagebus2::utils::PoolWorker>(10))
@@ -66,9 +66,10 @@ void AmqpClient::on_container_start(proton::container& container)
 {
     try {
         container.connect(m_url, connectOpts().reconnect(reconnectOpts()));
+        container.auto_stop(false);
     } catch (const std::exception& e) {
         logError("Exception {}", e.what());
-        m_connectPromise.set_value(ComState::ConnectFailed);
+        m_connectPromise.setValue(ComState::ConnectFailed);
     }
 }
 
@@ -80,42 +81,33 @@ void AmqpClient::on_container_stop(proton::container&)
 void AmqpClient::on_connection_open(proton::connection& connection)
 {
     m_connection = connection;
-
     if (connection.reconnected()) {
         logDebug("Reconnected on url: {}", m_url);
         resetPromises();
     } else {
         logDebug("Connected on url: {}", m_url);
     }
-    m_connectPromise.set_value(ComState::Connected);
+    m_connectPromise.setValue(ComState::Connected);
 }
 
 void AmqpClient::on_connection_close(proton::connection&)
 {
     logDebug("Close connection ...");
-    // TODO: To rework with common promise class (protection against promise already satisfied)
-    try {
-        m_deconnectPromise.set_value();
-    }
-    catch (const std::exception& e) {
-       logWarn("m_deconnectPromise error: {}", e.what());
-    }
+    m_deconnectPromise.setValue();
 }
 
 void AmqpClient::on_connection_error(proton::connection& connection)
 {
     logError("Error connection {}", connection.error().what());
-    // HOT_FIX: On connection error, the container is closed and the communication is definitively
+
+    // On connection error, the connection is closed and the communication is definitively
     // closed. In this particular case, the library never retry to reconnect the communication with
-    // reconnection option parameters.
-    // The idea is to propose a callback to execute when a communication error occurred to allow the
-    // client to restart the communication properly.
-    if (m_errorListener) {
-        resetPromises();
-        logInfo("Execute communication error callback");
-        // Execute callback treatment
-        m_pool->offload(std::bind(m_errorListener));
-    }
+    // reconnection option parameters. The idea is to try to re-open directly the connection.
+    // Note: The container must have the option auto_stop deactivated to not stop completely the container.
+    // Without that, the communication restoration will be more difficult.
+
+    m_connection = connection;
+    connection.work_queue().add([=]() { m_connection.open(); });
 }
 
 //void AmqpClient::on_connection_wake(proton::connection&)
@@ -136,8 +128,7 @@ void AmqpClient::on_sendable(proton::sender& sender)
         try {
             sender.send(m_message);
             sender.close();
-            // TODO: To rework with common promise class (protection against promise already satisfied)
-            m_promiseSender.set_value();
+            m_promiseSender.setValue();
             logDebug("Message sent");
         }
         catch (const std::exception& e) {
@@ -154,26 +145,13 @@ void AmqpClient::on_sender_close(proton::sender&)
 void AmqpClient::on_receiver_open(proton::receiver& receiver)
 {
     logDebug("Waiting any message on target address: {}", receiver.source().address());
-    // TODO: To rework with common promise class (protection against promise already satisfied)
-    try {
-        m_promiseReceiver.set_value();
-    }
-    catch (const std::exception& e) {
-       logWarn("m_promiseReceiver error: {}", e.what());
-    }
+    m_promiseReceiver.setValue();
 }
 
 void AmqpClient::on_receiver_close(proton::receiver&)
 {
     logDebug("Close receiver ...");
-    // TODO: Wait stop receiver doesn't work as expected: exception !!!!
-    //try {
-        // TODO: To rework with common promise class (protection against promise already satisfied)
-        //m_promiseReceiver.set_value();
-    //}
-    //catch (const std::exception& e) {
-    //   logWarn("m_promiseReceiver error: {}", e.what());
-    //}
+    m_promiseReceiver.setValue();
 }
 
 void AmqpClient::on_error(const proton::error_condition& error)
@@ -185,8 +163,13 @@ void AmqpClient::on_transport_error(proton::transport& transport)
 {
     logError("Transport error: {}", transport.error().what());
     // Reset connect promise in case of send or receive arrived before connection open
-    m_connectPromise = std::promise<fty::messagebus2::ComState>();
+    m_connectPromise.reset();
     m_communicationState = ComState::Lost;
+}
+
+void AmqpClient::on_transport_open(proton::transport&)
+{
+    logDebug("Open transport ...");
 }
 
 void AmqpClient::on_transport_close(proton::transport&)
@@ -198,39 +181,30 @@ void AmqpClient::resetPromises()
 {
     std::lock_guard<std::mutex> lock(m_lock);
     logDebug("Reset all promises");
-    m_connectPromise   = std::promise<fty::messagebus2::ComState>();
-    m_deconnectPromise = std::promise<void>();
-    m_promiseSender    = std::promise<void>();
-    m_promiseReceiver  = std::promise<void>();
+    m_connectPromise.reset();
+    m_deconnectPromise.reset();
+    m_promiseSender.reset();
+    m_promiseReceiver.reset();
 }
 
 bool AmqpClient::isConnected() {
-    // Wait communication was restored
+    // test if connected
     return (m_connection && m_connection.active() && connected() == ComState::Connected);
 }
 
 ComState AmqpClient::connected()
 {
-   //logTrace("AmqpClient::connected m_communicationState={}", m_communicationState);
+    // Wait communication was restored
     if (m_communicationState != ComState::Connected) {
-        auto connectFuture = m_connectPromise.get_future();
-        if (connectFuture.wait_for(TIMEOUT) != std::future_status::timeout) {
-            try {
-                m_communicationState = connectFuture.get();
-            } catch (const std::future_error& e) {
-                logError("Caught future error {}", e.what());
+        if (m_connectPromise.waitFor(TIMEOUT_MS)) {
+            if (auto value = m_connectPromise.getValue(); value) {
+                m_communicationState = *value;
             }
         } else {
             m_communicationState = ComState::ConnectFailed;
         }
     }
-    //logTrace("AmqpClient::connected m_communicationState={}", m_communicationState);
     return m_communicationState;
-}
-
-void AmqpClient::setConnectionErrorListener(ConnectionErrorListener errorListener)
-{
-    m_errorListener = errorListener;
 }
 
 DeliveryState AmqpClient::send(const proton::message& msg)
@@ -238,15 +212,15 @@ DeliveryState AmqpClient::send(const proton::message& msg)
     auto deliveryState = DeliveryState::Rejected;
     if (isConnected()) {
         std::lock_guard<std::mutex> lock(m_lockMain);
-        m_promiseSender = std::promise<void>();
+        m_promiseSender.reset();
         logDebug("Sending message to {} ...", msg.to());
         m_message.clear();
         m_message = msg;
         m_connection.work_queue().add([=]() {
-            m_connection.open_sender(msg.to());
+            m_connection.default_session().open_sender(msg.to());
         });
         // Wait to know if the message has been sent or not
-        if (m_promiseSender.get_future().wait_for(TIMEOUT) != std::future_status::timeout) {
+        if (m_promiseSender.waitFor(TIMEOUT_MS)) {
             deliveryState = DeliveryState::Accepted;
         }
     }
@@ -259,16 +233,20 @@ DeliveryState AmqpClient::receive(const Address& address, MessageListener messag
     if (isConnected()) {
         std::lock_guard<std::mutex> lock(m_lockMain);
         logDebug("Set receiver to wait message(s) from {} ...", address);
-        m_promiseReceiver = std::promise<void>();
+        m_promiseReceiver.reset();
 
-        (!filter.empty()) ? setSubscriptions(filter, messageListener) : setSubscriptions(address, messageListener);
+        auto name = setAddressFilter(address, filter);
+        setSubscriptions(name, messageListener);
 
         m_connection.work_queue().add([=]() {
-            m_connection.open_receiver(address, proton::receiver_options().auto_accept(true));
+            m_connection.default_session().open_receiver(address, proton::receiver_options().name(name).auto_accept(false));
         });
 
-        if (m_promiseReceiver.get_future().wait_for(TIMEOUT) != std::future_status::timeout) {
+        if (m_promiseReceiver.waitFor(TIMEOUT_MS)) {
             deliveryState = DeliveryState::Accepted;
+        }
+        else {
+            logError("Error on receive for {}, timeout reached", address);
         }
     }
     return deliveryState;
@@ -278,16 +256,13 @@ void AmqpClient::on_message(proton::delivery& delivery, proton::message& msg)
 {
     std::lock_guard<std::mutex> lock(m_lock);
     logDebug("Message arrived: {}", proton::to_string(msg));
-    delivery.accept();
     Message amqpMsg(getMetaData(msg), msg.body().empty() ? std::string{} : proton::to_string(msg.body()));
 
-    if (m_connection) {
-        std::string key = msg.address();
-        if (!msg.correlation_id().empty() && msg.reply_to().empty()) {
-            key = proton::to_string(msg.correlation_id());
-        }
-
+    if (m_connection.active()) {
+        std::string correlationId = msg.correlation_id().empty() || !msg.reply_to().empty() ? "" : proton::to_string(msg.correlation_id());
+        std::string key = setAddressFilter(msg.address(), correlationId);
         if (auto it {m_subscriptions.find(key)}; it != m_subscriptions.end()) {
+            delivery.accept();
             // Message listener called by qpid-proton library
 
             // NOTE: Need to execute the callback in another pool of thread. The callback
@@ -295,14 +270,17 @@ void AmqpClient::on_message(proton::delivery& delivery, proton::message& msg)
             // TODO: Don't work with another proton worker
             //m_workQueue.add(proton::make_work(it->second, amqpMsg));
             m_pool->offload(std::bind(it->second, amqpMsg));
+            return;
         }
         else {
             logWarn("No message listener checked in for: {}", key);
         }
     } else {
         // Connection not set
-        logError("Nothing to do, connection object not set");
+        logError("Nothing to do, connection object not active");
     }
+    logDebug("Message rejected: {}", proton::to_string(msg));
+    delivery.reject();
 }
 
 void AmqpClient::setSubscriptions(const Address& address, MessageListener messageListener)
@@ -338,52 +316,38 @@ void AmqpClient::unsetSubscriptions(const Address& address)
     }
 }
 
-DeliveryState AmqpClient::unreceive(const Address& address)
+DeliveryState AmqpClient::unreceive(const Address& address, const std::string& filter)
 {
     auto deliveryState = DeliveryState::Unavailable;
     if (isConnected()) {
         std::lock_guard<std::mutex> lock(m_lockMain);
 
-        // TODO: Remove filter doesn't work !!!
+        auto receiverName = setAddressFilter(address, filter);
         // Remove first address in subscriptions list
-        unsetSubscriptions(address);
+        unsetSubscriptions(receiverName);
 
         // Then find recever with input address and close it
         bool isFound = false;
-        auto receivers = m_connection.receivers();
+        auto receivers = m_connection.default_session().receivers();
+        logDebug("unreceive: try to close {}", receiverName);
         for (auto receiver : receivers) {
-            if (receiver.source().address() == address) {
+            if (receiver.name() == receiverName && !receiver.closed() && receiver.active()) {
                 isFound = true;
-                receiver.close();
-                deliveryState = DeliveryState::Accepted;
-                // TODO: Wait stop receiver dosn't work: exception !!!!
-                //m_connection.work_queue().add([&]() { receiver.close(); });
-                /*m_promiseReceiver = std::promise<void>();
-                m_connection.work_queue().add([&]() { receiver.close(); });
-                if (m_promiseReceiver.get_future().wait_for(TIMEOUT) != std::future_status::timeout) {
-                    logDebug("Receiver closed for {}", address);
+                m_promiseReceiver.reset();
+                m_connection.work_queue().add([&]() {
+                    receiver.close();
+                });
+                if (m_promiseReceiver.waitFor(TIMEOUT_MS)) {
+                    logDebug("Receiver closed for {}", receiverName);
                     deliveryState = DeliveryState::Accepted;
                 } else {
-                    logError("Error on unreceive for {}, timeout reached", address);
-                } */
+                    logError("Error on unreceive for {}, timeout reached", receiverName);
+                }
             }
         }
         if (!isFound)  {
-            logWarn("Unable to unreceive address {}: not present", address);
+            logWarn("Unable to unreceive address {}: not present", receiverName);
         }
-    }
-    return deliveryState;
-}
-
-DeliveryState AmqpClient::unreceiveFilter(const std::string& filter)
-{
-    auto deliveryState = DeliveryState::Unavailable;
-    if (isConnected()) {
-        std::lock_guard<std::mutex> lock(m_lockMain);
-
-        // Remove filter in subscriptions list
-        unsetSubscriptions(filter);
-        deliveryState = DeliveryState::Accepted;
     }
     return deliveryState;
 }
@@ -391,12 +355,11 @@ DeliveryState AmqpClient::unreceiveFilter(const std::string& filter)
 void AmqpClient::close()
 {
     std::lock_guard<std::mutex> lock(m_lock);
-    m_deconnectPromise = std::promise<void>();
+    m_deconnectPromise.reset();
     if (m_connection && m_connection.active()) {
         m_connection.work_queue().add([=]() { m_connection.close(); });
 
-        auto deconnectFuture = m_deconnectPromise.get_future();
-        if (deconnectFuture.wait_for(TIMEOUT) == std::future_status::timeout) {
+        if (!m_deconnectPromise.waitFor(TIMEOUT_MS)) {
             logError("AmqpClient::close De-connection timeout reached");
         }
     }