+++ amqp/src/AmqpClient.h
@@ -19,10 +19,10 @@
 
 #pragma once
 
+#include "fty/messagebus2/amqp/MessageBusAmqp.h"
 #include "MsgBusAmqpUtils.h"
-#include <fty/messagebus2/MessageBus.h>
-#include <fty/messagebus2/MessageBusStatus.h>
-#include <future>
+#include "fty/messagebus2/utils/MsgBusPoolWorker.hpp"
+#include "fty/messagebus2/Promise.h"
 #include <proton/connection.hpp>
 #include <proton/container.hpp>
 #include <proton/delivery.hpp>
@@ -34,12 +34,12 @@
 #include <proton/transport.hpp>
 #include <proton/work_queue.hpp>
 
-#include "fty/messagebus2/utils/MsgBusPoolWorker.hpp"
-
 namespace fty::messagebus2::amqp {
 
 using MessageListener      = fty::messagebus2::MessageListener;
 using SubScriptionListener = std::map<Address, MessageListener>;
+class AmqpClient;
+using AmqpClientPointer = std::shared_ptr<AmqpClient>;
 
 class AmqpClient : public proton::messaging_handler
 {
@@ -63,28 +63,26 @@ public:
     void on_error(const proton::error_condition& error) override;
     void on_transport_error(proton::transport& t) override;
     void on_transport_close(proton::transport&) override;
+    void on_transport_open(proton::transport&) override;
 
     fty::messagebus2::ComState connected();
-    bool isConnected();
-    void setConnectionErrorListener(ConnectionErrorListener errorListener = {});
+    bool isConnected();    
     fty::messagebus2::DeliveryState receive(
         const Address& address, MessageListener messageListener = {}, const std::string& filter = {});
-    fty::messagebus2::DeliveryState unreceive(const Address& address);
-    fty::messagebus2::DeliveryState unreceiveFilter(const std::string& filter);
+    fty::messagebus2::DeliveryState unreceive(const Address& address, const std::string& filter = {});
     fty::messagebus2::DeliveryState send(const proton::message& msg);
     void close();
 
 private:
     Endpoint                m_url;
-    SubScriptionListener    m_subscriptions;
-    ConnectionErrorListener m_errorListener;
+    SubScriptionListener    m_subscriptions;    
 
     // Default communication state
     fty::messagebus2::ComState m_communicationState = fty::messagebus2::ComState::Unknown;
 
     // Proton object
-    proton::connection   m_connection;
-    proton::message      m_message;
+    proton::connection    m_connection;
+    proton::message       m_message;
 
     // Pool thread
     std::shared_ptr<fty::messagebus2::utils::PoolWorker> m_pool;
@@ -93,13 +91,13 @@ private:
     // TODO: Refactoring mutex mgt in application ...
     std::mutex m_lock;
     std::mutex m_lockMain;
-
+    
     // Set of promise for synchronization
-    std::promise<fty::messagebus2::ComState> m_connectPromise;
-    std::promise<void>                       m_deconnectPromise;
-    std::promise<void>                       m_promiseSender;
-    std::promise<void>                       m_promiseReceiver;
-    std::promise<void>                       m_promiseSenderClose;
+    Promise<fty::messagebus2::ComState> m_connectPromise;
+    Promise<void>                       m_deconnectPromise;
+    Promise<void>                       m_promiseSender;
+    Promise<void>                       m_promiseReceiver;
+    Promise<void>                       m_promiseSenderClose;
 
     void setSubscriptions(const Address& address, MessageListener messageListener);
     void unsetSubscriptions(const Address& address);