+++ amqp/src/MsgBusAmqpUtils.h
@@ -29,6 +29,25 @@ namespace fty::messagebus2::amqp {
 
 using property_map = std::map<std::string, proton::scalar>;
 
+inline std::string setAddressFilter(const Address& address, const std::string& filter={})
+{
+    return filter.empty() ? address : address + "|" + filter;
+}
+
+inline std::pair<std::string, std::string> getAddressFilter(const std::string& input)
+{
+    std::pair<std::string, std::string> ret;
+    if (auto pos = input.find("|"); pos != std::string::npos) {
+        ret.first = input.substr(0, pos);
+        ret.second = input.substr(pos + 1);
+    }
+    else {
+        ret.first = input;
+        ret.second = "";
+    }
+    return ret;
+}
+
 inline const MetaData getMetaData(const proton::message& protonMsg)
 {
     Message message;